Test whether a string is a valid UKFSR indicator id

A UKFSR indicator can be uniquely identified from 3 main hierarchical components separated by periods:

theme (1-5) but see Food Security Index below

section (1-9)

indicator (1-99)

with an optional variant suffix in the form (a-z). Thus a valid indicator id is a string in the form "T.S.Iv". Some examples are "1.3.7a". "4.2.11". check_indicator tests an indicator id string for compliance and returns TRUE or FALSE.

Food Security Index

The FSI also contains a set of indicators. For FSI2024 we used the UKFSR infrastructure to generate the graphics. As a result there is an exceptional set of indicator ids possible, where the theme can be 'fsi'.

Usage

check_indicator(string)

Arguments

string

String containing a UKFSR indicator id

Value

A logical

See also

Examples

# A valid id
check_indicator("2.3.5a")
#> [1] TRUE

# An invalid id
check_indicator("10.3.1")
#> [1] FALSE